#18 Refactor: Making a component for loading overlay, and refactoring API response
