+++ code transpiler react app/src/Components/CommonComponents/LinkComponent.tsx
