+++ code transpiler react app/src/Components/CommonComponents/LoadingOverlay/LoadingOverlay.module.css
@@ -1,0 +1,12 @@
+.overlay {
+    display: flex;
+    width:100vw;
+    height:100vh;
+    align-items:center;
+    justify-content:center;
+    position:fixed;
+    background-color:rgba(0,0,0,0.5);
+    z-index:6;
+    margin-left: -32px;
+    margin-top: -32px;
+}
